removed class RefU16Ret

+++ src/main/java/org/gutkyu/dosboxj/dos/mem_block/DOSPSP.java
@@ -92,9 +92,9 @@ public final class DOSPSP extends MemStruct {
         }
     }
 
-    public short findFreeFileEntry() {
+    public int findFreeFileEntry() {
         int files = Memory.real2Phys(getIt(Size_file_table, Off_file_table));
-        for (short i = 0; i < getIt(Size_max_files, Off_max_files); i++) {
+        for (int i = 0; i < getIt(Size_max_files, Off_max_files); i++) {
             if (Memory.readB(files + i) == 0xff)
                 return i;
         }
